+++ src/route/ExamRoutes.ts
@@ -1,5 +1,5 @@
 import {ExamController} from "../controller/ExamController";
-import {Config as config} from "./config/Config";
+import {Config as config} from "./Config";
 
 const BASE_PATH = "/exams";
