+++ src/route/NoteRoutes.ts
@@ -1,5 +1,5 @@
 import {NoteController} from "../controller/NoteController";
-import {Config as config} from "./config/Config";
+import {Config as config} from "./Config";
 
 const BASE_PATH = "/notes";
