+++ src/route/UserRoutes.ts
@@ -1,5 +1,5 @@
 import {UserController} from "../controller/UserController";
-import {Config as config} from "./config/Config";
+import {Config as config} from "./Config";
 
 const BASE_PATH = "/users";
